+++ src/components/common/content/ContentCard.jsx
